rename and fixup
