--- src/ndv/controller/__init__.py
@@ -1,3 +1,0 @@
-from ._controller import ArrayViewer
-
-__all__ = ["ArrayViewer"]
