+++ src/ndv/controllers/__init__.py
@@ -1,0 +1,3 @@
+from ._array_viewer import ArrayViewer
+
+__all__ = ["ArrayViewer"]
